+++ base/gmp.jl
@@ -348,15 +348,16 @@ end
 
 rem(x::BigInt, ::Type{Bool}) = !iszero(x) & unsafe_load(x.d) % Bool # never unsafe here
 
-rem(x::BigInt, ::Type{T}) where T<:Union{SLimbMax,ULimbMax} =
-    iszero(x) ? zero(T) : flipsign(unsafe_load(x.d) % T, x.size)
-
-function rem(x::BigInt, ::Type{T}) where T<:Union{Base.BitUnsigned,Base.BitSigned}
-    u = zero(T)
-    for l = 1:min(abs(x.size), cld(sizeof(T), sizeof(Limb)))
-        u += (unsafe_load(x.d, l) % T) << ((sizeof(Limb)<<3)*(l-1))
+function rem(x::BigInt, ::Type{T}) where T<:Base.BitInteger
+    if sizeof(T) <= sizeof(Limb)
+        iszero(x) ? zero(T) : flipsign(unsafe_load(x.d) % T, x.size)
+    else
+        u = zero(T)
+        for l = 1:min(abs(x.size), cld(sizeof(T), sizeof(Limb)))
+            u += (unsafe_load(x.d, l) % T) << ((sizeof(Limb)<<3)*(l-1))
+        end
+        flipsign(u, x.size)
     end
-    flipsign(u, x.size)
 end
 
 rem(x::Integer, ::Type{BigInt}) = BigInt(x)