+++ base/strings/string.jl
@@ -556,7 +556,7 @@ isvalid(s::String, i::Int) = checkbounds(Bool, s, i) && thisind(s, i) == i
 isascii(s::String) = isascii(codeunits(s))
 
 # don't assume effects for general integers since we cannot know their implementation
-@assume_effects :foldable repeat(c::Char, r::BitInteger) = @invoke repeat(c::Char, r::Integer)
+@assume_effects :foldable repeat(c::Char, r::BitInteger128) = @invoke repeat(c::Char, r::Integer)
 
 """
     repeat(c::AbstractChar, r::Integer)::String