+++ stdlib/Random/src/MersenneTwister.jl
@@ -489,7 +489,7 @@ function rand!(r::MersenneTwister, A::UnsafeView{UInt128}, ::SamplerType{UInt128
     A
 end
 
-for T in BitInteger_types
+for T in BitInteger128_types
     @eval function rand!(r::MersenneTwister, A::Array{$T}, sp::SamplerType{$T})
         GC.@preserve A rand!(r, UnsafeView(pointer(A), length(A)), sp)
         A