+++ stdlib/Random/src/Random.jl
@@ -15,7 +15,7 @@ using Base.GMP.MPZ
 using Base.GMP: Limb
 using SHA: SHA, SHA2_256_CTX, SHA2_512_CTX, SHA_CTX
 
-using Base: BitInteger, BitInteger_types, BitUnsigned, require_one_based_indexing
+using Base: BitInteger, BitInteger128_types, BitUnsigned, require_one_based_indexing
 import Base: copymutable, copy, copy!, ==, hash, convert,
              rand, randn, show
 
@@ -167,7 +167,7 @@ Base.getindex(::SamplerType{T}) where {T} = T
 
 # SamplerUnion(X, Y, ...}) == Union{SamplerType{X}, SamplerType{Y}, ...}
 SamplerUnion(U...) = Union{Any[SamplerType{T} for T in U]...}
-const SamplerBoolBitInteger = SamplerUnion(Bool, BitInteger_types...)
+const SamplerBoolBitInteger = SamplerUnion(Bool, BitInteger128_types...)
 
 
 struct SamplerTrivial{T,E} <: Sampler{E}