+++ NEWS.md
@@ -11,6 +11,10 @@ Language changes
 * `mod(x::AbstractFloat, -Inf)` now returns `x` (as long as `x` is finite), this aligns with C standard and
 is considered a bug fix ([#47102])
 
+  - `Base` internal types `BitSigned` and `BitUnsigned` now refer to abstract types instead of unions, and
+    `BitInteger` is defined to `Union{BitSigned, BitUnsigned}`; builtin integer types inherit either from
+    `BitSigned` or `BitUnsigned`.
+
 Compiler/Runtime improvements
 -----------------------------